+++ src/KurrentDB.Core/Certificates/DevCertificateProvider.cs
@@ -10,6 +10,8 @@ public class DevCertificateProvider : CertificateProvider {
 	public DevCertificateProvider(X509Certificate2 certificate) {
 		Certificate = certificate;
 		TrustedRootCerts = new X509Certificate2Collection(certificate);
+		NodeClientCertificate = certificate;
+		NodeClientTrustedRootCerts = TrustedRootCerts;
 	}
 	public override LoadCertificateResult LoadCertificates(ClusterVNodeOptions options) {
 		return LoadCertificateResult.Skipped;