+++ src/KurrentDB.Core/Certificates/OptionsCertificateProvider.cs
@@ -9,6 +9,7 @@ using Serilog;
 
 namespace KurrentDB.Core.Certificates;
 
+// (Re)Loads the certificates specified in the ClusterVNodeOptions
 public class OptionsCertificateProvider : CertificateProvider {
 	private static readonly ILogger Log = Serilog.Log.ForContext<ClusterVNode>();
 	private string _cachedReservedNodeCN;
@@ -19,56 +20,105 @@ public class OptionsCertificateProvider : CertificateProvider {
 			return LoadCertificateResult.Skipped;
 		}
 
+		// Load both Node and NodeClient certificates up-front.
+		// Node cert is the server cert sent to incoming connections
+		// NodeClient cert is the client cert sent on outgoing connections to other nodes.
+		// NodeClient cert defaults to the same certificate as Node cert if not provided.
 		var (certificate, intermediates) = options.LoadNodeCertificate();
+		var hasNodeClientCert = options.TryLoadNodeClientCertificate(out var nodeClientCertificate, out var nodeClientIntermediates);
+		if (!hasNodeClientCert) {
+			nodeClientCertificate = certificate;
+			nodeClientIntermediates = intermediates;
+		}
 
 		string reservedNodeCN;
 		var reservedNodeCNOption = nameof(options.Certificate.CertificateReservedNodeCommonName);
 
+		// Determine the CN pattern expected from incoming node certificates.
 		if (options.Certificate.CertificateReservedNodeCommonName.IsNotEmptyString()) {
+			// Reserved CN is configured. Check that the node client cert matches it.
 			reservedNodeCN = options.Certificate.CertificateReservedNodeCommonName;
-			if (!certificate.ClientCertificateMatchesName(reservedNodeCN)) {
-				var certificateCN = certificate.GetCommonName();
+			if (!nodeClientCertificate.ClientCertificateMatchesName(reservedNodeCN)) {
+				var nodeClientCertCN = nodeClientCertificate.GetCommonName();
 				Log.Error(
 					"Certificate CN: {certificateCN} does not match with the {reservedNodeCNOption} configuration setting: {reservedNodeCN}",
-					certificateCN, reservedNodeCNOption, reservedNodeCN);
+					nodeClientCertCN, reservedNodeCNOption, reservedNodeCN);
 				return LoadCertificateResult.VerificationFailed;
 			}
 			Log.Information("{reservedNodeCNOption} configured to: {reservedNodeCN}",
 				reservedNodeCNOption, reservedNodeCN);
 		} else {
-			reservedNodeCN = certificate.GetCommonName();
+			reservedNodeCN = nodeClientCertificate.GetCommonName();
 			Log.Information("{reservedNodeCNOption} auto-configured to: {reservedNodeCN} based on certificate",
 				reservedNodeCNOption, reservedNodeCN);
 		}
 
-		var previousThumbprint = Certificate?.Thumbprint;
-		var newThumbprint = certificate.Thumbprint;
-		Log.Information("Loading the node's certificate. Subject: {subject}, Previous thumbprint: {previousThumbprint}, New thumbprint: {newThumbprint}",
-			certificate.SubjectName.Name, previousThumbprint, newThumbprint);
+		// Log information about the certificates and their intermediates
+		LogThumbprints("node", Certificate, certificate, intermediates);
+		LogThumbprints("node client", NodeClientCertificate, nodeClientCertificate, nodeClientIntermediates);
 
-		if (intermediates != null) {
-			foreach (var intermediateCert in intermediates) {
-				Log.Information("Loading intermediate certificate. Subject: {subject}, Thumbprint: {thumbprint}", intermediateCert.SubjectName.Name, intermediateCert.Thumbprint);
+		static void LogThumbprints(string label, X509Certificate2 previousCertificate, X509Certificate2 certificate, X509Certificate2Collection intermediates) {
+			var previousThumbprint = previousCertificate?.Thumbprint;
+			var newThumbprint = certificate.Thumbprint;
+			Log.Information("Loading the {label} certificate. Subject: {subject}, Previous thumbprint: {previousThumbprint}, New thumbprint: {newThumbprint}",
+				label, certificate.SubjectName.Name, previousThumbprint, newThumbprint);
+
+			if (intermediates != null) {
+				foreach (var intermediateCert in intermediates) {
+					Log.Information("Loading {label} intermediate certificate. Subject: {subject}, Thumbprint: {thumbprint}",
+						label, intermediateCert.SubjectName.Name, intermediateCert.Thumbprint);
+				}
 			}
 		}
 
+		// Validate the node certificate
 		var trustedRootCerts = options.LoadTrustedRootCertificates();
 
 		foreach (var trustedRootCert in trustedRootCerts) {
-			Log.Information("Loading trusted root certificate. Subject: {subject}, Thumbprint: {thumbprint}", trustedRootCert.SubjectName.Name, trustedRootCert.Thumbprint);
+			Log.Information("Loading trusted root for node certificate. Subject: {subject}, Thumbprint: {thumbprint}", trustedRootCert.SubjectName.Name, trustedRootCert.Thumbprint);
 		}
 
-		if (!VerifyCertificates(certificate, intermediates, trustedRootCerts)) {
+		if (!VerifyCertificates("node", certificate, intermediates, trustedRootCerts)) {
 			return LoadCertificateResult.VerificationFailed;
 		}
 
+		// Validate the node client certificate
+		var nodeClientTrustedRootCerts = hasNodeClientCert
+			? options.LoadNodeClientTrustedRootCertificates()
+			: trustedRootCerts;
+
+		if (hasNodeClientCert) {
+			foreach (var trustedRootCert in nodeClientTrustedRootCerts) {
+				Log.Information("Loading trusted root for node client certificate. Subject: {subject}, Thumbprint: {thumbprint}", trustedRootCert.SubjectName.Name, trustedRootCert.Thumbprint);
+			}
+
+			if (!VerifyCertificates("node client", nodeClientCertificate, nodeClientIntermediates, nodeClientTrustedRootCerts)) {
+				return LoadCertificateResult.VerificationFailed;
+			}
+		}
+
+		// Check our client certificate will be classified correctly by other nodes
+		if (options.Cluster.ClusterSize > 1 && nodeClientCertificate.ClassifyInboundCertificate(
+				disableClientAuthEkuValidation: options.Certificate.DisableClientAuthEkuValidation,
+				out var nodeClientCertDescription) is not CertificateClassification.Node) {
+
+			Log.Error(hasNodeClientCert
+				? "The node client certificate was not recognized as a node certificate: {description}"
+				: "The node certificate was not recognized as a node certificate: {description}",
+				nodeClientCertDescription);
+			return LoadCertificateResult.VerificationFailed;
+		}
+	
 		// no need for a lock here since reference assignment is atomic. however, other threads may not immediately
 		// see the changes and the order in which they see the changes is also not guaranteed as we don't have any
 		// memory barriers here. this is not a problem as in the worst case, it will cause the certificate verifications
 		// to fail when establishing/receiving a connection and the next connection retries will succeed.
 		Certificate = certificate;
 		IntermediateCerts = intermediates;
+		NodeClientCertificate = nodeClientCertificate;
+		NodeClientIntermediateCerts = nodeClientIntermediates;
 		TrustedRootCerts = trustedRootCerts;
+		NodeClientTrustedRootCerts = nodeClientTrustedRootCerts;
 		_cachedReservedNodeCN = reservedNodeCN;
 
 		Log.Information("All certificates successfully loaded.");
@@ -79,18 +129,23 @@ public class OptionsCertificateProvider : CertificateProvider {
 		return _cachedReservedNodeCN ?? throw new InvalidOperationException("Certificates are not loaded.");
 	}
 
-	private static bool VerifyCertificates(X509Certificate2 nodeCertificate, X509Certificate2Collection intermediates, X509Certificate2Collection trustedRoots) {
+	private static bool VerifyCertificates(
+		string label,
+		X509Certificate2 certificate,
+		X509Certificate2Collection intermediates,
+		X509Certificate2Collection trustedRoots) {
+
 		bool error = false;
 
-		if (!CertificateUtils.IsValidNodeCertificate(nodeCertificate, out var errorMsg)) {
-			Log.Error(errorMsg);
+		if (!CertificateUtils.IsValidNodeCertificate(certificate, out var errorMsg)) {
+			Log.Error("The {label} certificate: {error}", label, errorMsg);
 			error = true;
 		}
 
 		if (intermediates != null) {
 			foreach (var cert in intermediates) {
 				if (!CertificateUtils.IsValidIntermediateCertificate(cert, out errorMsg)) {
-					Log.Error($"{errorMsg} Please bundle only intermediate certificates (if any) and not root certificates with the node's certificate.");
+					Log.Error("{error} Please bundle only intermediate certificates (if any) and not root certificates with the {label} certificate.", errorMsg, label);
 					error = true;
 				}
 			}
@@ -99,24 +154,25 @@ public class OptionsCertificateProvider : CertificateProvider {
 		if (trustedRoots != null && trustedRoots.Count > 0) {
 			foreach (var cert in trustedRoots) {
 				if (!CertificateUtils.IsValidRootCertificate(cert, out errorMsg)) {
-					Log.Error($"{errorMsg} If you have intermediate certificates, please bundle them with the node's certificate (in PEM or PKCS #12 format).");
+					Log.Error("{error} If you have intermediate certificates, please bundle them with the {label} certificate (in PEM or PKCS #12 format).", errorMsg, label);
 					error = true;
 				}
 			}
 		} else {
-			Log.Error("No trusted root certificates loaded");
+			Log.Error("No trusted root certificates loaded for the {label} certificate", label);
 			error = true;
 		}
 
 		if (error)
 			return false;
 
-		var chainStatus = CertificateUtils.BuildChain(nodeCertificate, intermediates, trustedRoots, out var chainStatusInformation);
+		var chainStatus = CertificateUtils.BuildChain(certificate, intermediates, trustedRoots, out var chainStatusInformation);
 
 		if (chainStatus != X509ChainStatusFlags.NoError) {
 			Log.Error(
-				"Failed to build the certificate chain with the node's own certificate up to the root. " +
-				"If you have intermediate certificates, please bundle them with the node's certificate (in PEM or PKCS #12 format). Errors:-");
+				"Failed to build the certificate chain with the {label} certificate up to the root. " +
+				"If you have intermediate certificates, please bundle them with the {label} certificate (in PEM or PKCS #12 format). Errors:-",
+				label, label);
 			foreach (var status in chainStatusInformation) {
 				Log.Error(status);
 			}
@@ -125,7 +181,7 @@ public class OptionsCertificateProvider : CertificateProvider {
 		}
 
 		if (!error && intermediates != null) {
-			chainStatus = CertificateUtils.BuildChain(nodeCertificate, null, trustedRoots, out chainStatusInformation);
+			chainStatus = CertificateUtils.BuildChain(certificate, null, trustedRoots, out chainStatusInformation);
 
 			// Adding the intermediate certificates to the store is required so that
 			// i)  the full certificate chain (excluding the root) is sent from client to server (on both Windows/Linux)
@@ -146,7 +202,7 @@ public class OptionsCertificateProvider : CertificateProvider {
 		}
 
 		if (!error) {
-			Log.Information("Certificate chain verification successful.");
+			Log.Information("The {label} certificate chain verification successful.", label);
 		}
 
 		return !error;